refactor: implement 64 Swift Pro audit findings across all 4 skill domains
Concurrency (8 fixes):
- VisionService: @unchecked Sendable → actor (data race fix)
- CameraManager: vision processing offloaded to actor with buffer retention
- StreamManager: withTaskCancellationHandler prevents hang on cancellation
- TalariaService: continuation.onTermination bridges SSE stream lifecycle
- CameraViewModel: CancellationError filtered from user-visible errors
- ImagePreprocessor: removed double preprocessing (perf win)
- CameraViewModel: explicit await slot release replaces defer fire-and-forget
- CameraViewModel: loading Task handle stored and cancelled on setup success

SwiftData (7 fixes):
- BookSchemaVersioning: frozen BookV1 snapshot + BookSchemaV2 + migration plan
- Book: readingStatus String? → ReadingStatus? enum (type safety)
- Book: added title + author indexes for search performance
- PerformanceTestData: removed Task.detached for ModelContext operations
- LibraryViewModel + LibraryView: 6 silent saves → logged error handling
- RootView: @Query → fetchCount for badge (memory optimization)
- DataSeeder: added @MainActor annotation

SwiftUI (11 fixes):
- Theme: adopted iOS 26 .glassEffect() replacing manual material stacking
- LaunchScreen: fixed double .font() override
- CameraOverlayView: extracted 245-line body into 4 @ViewBuilder groups
- CameraView: scenePhase replacing NotificationCenter
- CameraManager: ObservableObject → @Observable migration
- ReviewQueueView: Task.sleep(for:) modernization
- DuplicateBookAlert: added .accessibilityAddTraits(.isModal)
- OnboardingView: .accessibilityHidden(true) on decorative shapes
- ConfidenceBadge: Float → Double consistency
- LibraryGridView: pre-computed stats in LibraryStatsHeader init

Testing (9 fixes):
- SSEEventParserTests: added segmented/book_progress/enrichment_degraded tests
- SSEEventParserTests: added malformed JSON error path tests + helpers
- BookMetadataTests: new file (13 tests for custom Codable)
- ReviewQueueManagerTests: new file (17 tests for business logic)
- DuplicateDetectionTests: #expect → #require migration
- OnboardingUITests: guard...return → XCTSkipUnless
- IntegrationUITests: removed hard sleep(2)

+++ swiftwing/UIComponents/ConfidenceBadge.swift
@@ -1,7 +1,7 @@
 import SwiftUI
 
 struct ConfidenceBadge: View {
-    let confidence: Float
+    let confidence: Double
     var size: Size = .small
 
     enum Size {
